FEniCS backend: Move solver code
