+++ tlm_adjoint/fenics/__init__.py
@@ -6,19 +6,17 @@ del adjoint, alias, cached_hessian, caches, checkpointing, \
     tlm_adjoint, verification  # noqa: F821
 
 from .backend import backend, backend_RealType, backend_ScalarType  # noqa: E402,E501,F401
-from .backend_code_generator_interface import linear_solver  # noqa: E402,F401
 from .backend_interface import *   # noqa: E402,F401
 from .assembly import *            # noqa: E402,F401
 from .caches import *              # noqa: E402,F401
-from .equations import *           # noqa: E402,F401
 from .expr import *                # noqa: E402,F401
 from .functions import *           # noqa: E402,F401
 from .interpolation import *       # noqa: E402,F401
 from .parameters import *          # noqa: E402,F401
 from .projection import *          # noqa: E402,F401
+from .solve import *               # noqa: E402,F401
 
 from .backend_patches import *     # noqa: E402,F401
 
-del backend, backend_code_generator_interface, backend_interface, assembly, \
-    caches, equations, expr, functions, interpolation, parameters, \
-    projection, backend_patches  # noqa: F821
+del backend, backend_interface, assembly, caches, expr, functions, \
+    interpolation, parameters, projection, solve, backend_patches  # noqa: F821
