--- tlm_adjoint/fenics/backend_code_generator_interface.py
@@ -1,84 +1,0 @@
-from .backend import (
-    LUSolver, KrylovSolver, Parameters, backend_DirichletBC, backend_solve as
-    solve, has_lu_solver_method, parameters)
-from ..interface import DEFAULT_COMM
-
-from .assembly import assemble_matrix, assemble_system
-from .parameters import update_parameters
-
-__all__ = \
-    [
-        "linear_solver",
-
-        "solve"
-    ]
-
-
-if "tlm_adjoint" not in parameters:
-    parameters.add(Parameters("tlm_adjoint"))
-_parameters = parameters["tlm_adjoint"]
-if "EquationSolver" not in _parameters:
-    _parameters.add(Parameters("EquationSolver"))
-if "enable_jacobian_caching" not in _parameters["EquationSolver"]:
-    _parameters["EquationSolver"].add("enable_jacobian_caching", True)
-if "cache_rhs_assembly" not in _parameters["EquationSolver"]:
-    _parameters["EquationSolver"].add("cache_rhs_assembly", True)
-if "match_quadrature" not in _parameters["EquationSolver"]:
-    _parameters["EquationSolver"].add("match_quadrature", False)
-del _parameters
-
-
-def assemble_linear_solver(A_form, b_form=None, bcs=None, *,
-                           form_compiler_parameters=None,
-                           linear_solver_parameters=None):
-    if bcs is None:
-        bcs = ()
-    elif isinstance(bcs, backend_DirichletBC):
-        bcs = (bcs,)
-    if form_compiler_parameters is None:
-        form_compiler_parameters = {}
-    if linear_solver_parameters is None:
-        linear_solver_parameters = {}
-
-    if b_form is None:
-        A, b = assemble_matrix(
-            A_form, bcs=bcs, form_compiler_parameters=form_compiler_parameters)
-    else:
-        A, b = assemble_system(
-            A_form, b_form, bcs=bcs,
-            form_compiler_parameters=form_compiler_parameters)
-
-    solver = linear_solver(A, linear_solver_parameters)
-
-    return solver, A, b
-
-
-def linear_solver(A, linear_solver_parameters, *, comm=None):
-    if comm is None:
-        if hasattr(A, "mpi_comm"):
-            comm = A.mpi_comm()
-        else:
-            comm = DEFAULT_COMM
-
-    linear_solver = linear_solver_parameters.get("linear_solver", "default")
-    if linear_solver in {"direct", "lu"}:
-        linear_solver = "default"
-    elif linear_solver == "iterative":
-        linear_solver = "gmres"
-    is_lu_linear_solver = linear_solver == "default" \
-        or has_lu_solver_method(linear_solver)
-    if is_lu_linear_solver:
-        solver = LUSolver(comm, linear_solver)
-        solver.set_operator(A)
-        lu_parameters = linear_solver_parameters.get("lu_solver", {})
-        update_parameters(solver.parameters, lu_parameters)
-    else:
-        pc = linear_solver_parameters.get("preconditioner", "default")
-        ks_parameters = linear_solver_parameters.get("krylov_solver", {})
-        solver = KrylovSolver(comm, linear_solver, pc)
-        solver.set_operator(A)
-        update_parameters(solver.parameters, ks_parameters)
-    return solver
-
-
-# def solve(*args, **kwargs):
